app/requirements.txt
amqp==5.1.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.7.2
astroid==2.4.2
async-timeout==4.0.2
autopep8==1.5.4
beautifulsoup4==4.10.0
billiard==4.2.0
boto3==1.24.36
botocore==1.27.36
Brotli==1.0.9
celery==5.4.0
celery-redbeat==2.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextlib2==21.6.0
crispy-bootstrap5==2023.10
cryptography==3.3.1
cssselect2==0.4.1
defusedxml==0.6.0
Deprecated==1.2.13
deprecation==2.1.0
Django==4.2.7
django-allauth==0.50.0
django-crispy-forms==2.1
django-model-utils==5.0.0
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
django-notifications-hq==1.8.3
django-scheduler==0.10.1
django-storages==1.12.3
django-tenants==3.5.0
fonttools==4.29.0
future==0.18.2
html5lib==1.1
icalendar==6.0.1
idna==2.10
injector==0.22.0
isort==5.7.0
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
kombu==5.3.7
lazy-object-proxy==1.4.3
lxml==4.9.3
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.1.0
ordered-set==4.1.0
packaging==21.3
pdf2image==1.16.0
pikepdf==8.6.0
Pillow==10.1.0
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
pycodestyle==2.6.0
pycparser==2.20
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.2.0
pyetherpadlite==1.1.1
PyJWT==2.0.1
pylint==2.6.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyphen==0.12.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.1
pyUFbr==0.1.0
redis==4.3.3
reportlab==4.0.6
requests==2.28.1
requests-oauthlib==1.3.0
s3transfer==0.6.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.4.0
tenacity==9.0.0
tenant-schemas-celery==2.0.0
tesserocr==2.7.0
tinycss2==1.1.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.3
validate-docbr==1.10.0
vine==5.1.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.12.1
zopfli==0.1.9
